FlaskProject/requirements.txt
Flask==2.0.2
Flask-RESTful==0.3.8
PyJWT==2.3.0
Flask-SQLAlchemy==2.5
PyQuery==1.2.4